Made changes in backend PolicyDTOProtoUtil for adding validation for new field and corrected compactRepr; Added input field in policyComponent UI; Removed unused PolicyDTOProtoUtil from userapi;
